+++ Genesis/Changes.md
@@ -7,6 +7,7 @@ new mutable mouse event handlers.
 - Path no longer extends DistanceLike
 - Renamed RGB, RGBLike and RGBChannel to Rgb RgbLike and RgbChannel
 - Renamed HSL and HSLLike to Hsl and HslLike
+- KeyTypedEvents now contain the associated key index
 ### New Features
 - Added global mouse and keyboard event handling with GlobalMouseEventHandler and GlobalKeyboardEventHandler objects
 - Added SegmentedPath for path / animation based on a set of values